requirement.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
joblib==0.15.1
pycryptodome==3.9.8
pycryptodomex==3.9.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
threadpoolctl==2.1.0
urllib3==1.25.10